Big boy commit - read description
Added shooter prefab, and made it so that the shooty boys don't fire while in air or in bullet time. They also don't immediately shoot after spawning in.  Next step is to have a check for distance from player.

The reason every single scene has been update is stupid and my fault. I was calculating the launch angle for the enemies using the wrong transform. I should have been using the player transform for everything.